去掉之江定制插接箱零线电流

+++ common/cfgcom/config.h
@@ -76,7 +76,7 @@ struct sObjCfg
 
 struct sIpCfg {
     QString addr;
-    uchar version;
+    ushort version;
     uchar language; //
     uchar lines;
     uchar modbus; //  1 表示 Modbus版本
